refactor: move the RhIdentity extraction to SecurityContextUtil class
It makes more sense to have the RhIdentity extraction method in the
SecurityContextUtil class because the latter already has many methods
that work with the security context itself.

RHCLOUD-34485
