+++ backend/src/main/java/com/redhat/cloud/notifications/routers/SecurityContextUtil.java
@@ -1,7 +1,10 @@
 package com.redhat.cloud.notifications.routers;
 
 import com.redhat.cloud.notifications.auth.ConsoleIdentityProvider;
+import com.redhat.cloud.notifications.auth.principal.ConsoleIdentity;
+import com.redhat.cloud.notifications.auth.principal.ConsolePrincipal;
 import com.redhat.cloud.notifications.auth.principal.rhid.RhIdPrincipal;
+import com.redhat.cloud.notifications.auth.principal.rhid.RhIdentity;
 import com.redhat.cloud.notifications.db.repositories.ApplicationRepository;
 import com.redhat.cloud.notifications.db.repositories.InternalRoleAccessRepository;
 import com.redhat.cloud.notifications.models.InternalRoleAccess;
@@ -10,6 +13,7 @@ import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.core.SecurityContext;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.UUID;
 
@@ -39,6 +43,25 @@ public class SecurityContextUtil {
 
     public static Boolean isServiceAccountAuthentication(SecurityContext securityContext) {
         return "ServiceAccount".equals(((RhIdPrincipal) securityContext.getUserPrincipal()).getType());
+    }
+
+    /**
+     * Extracts the {@link RhIdentity} object from the security context.
+     * @param securityContext the security context to extract the object from.
+     * @return the extracted {@link RhIdentity} object.
+     */
+    public static RhIdentity extractRhIdentity(final SecurityContext securityContext) {
+        final Principal genericPrincipal = securityContext.getUserPrincipal();
+        if (!(genericPrincipal instanceof ConsolePrincipal<?> principal)) {
+            throw new IllegalStateException(String.format("unable to extract RH Identity object from principal. Expected \"Console Principal\" object type, got \"%s\"", genericPrincipal.getClass().getName()));
+        }
+
+        final ConsoleIdentity genericIdentity = principal.getIdentity();
+        if (!(genericIdentity instanceof RhIdentity identity)) {
+            throw new IllegalStateException(String.format("unable to extract RH Identity object from principal. Expected \"RhIdentity\" object type, got \"%s\"", genericIdentity.getClass().getName()));
+        }
+
+        return identity;
     }
 
     public void hasPermissionForRole(SecurityContext securityContext, String role) {
