+++ backend/src/test/java/com/redhat/cloud/notifications/auth/kessel/KesselAuthorizationTest.java
@@ -8,8 +8,6 @@ import com.redhat.cloud.notifications.auth.principal.rhid.RhIdPrincipal;
 import com.redhat.cloud.notifications.auth.principal.rhid.RhIdentity;
 import com.redhat.cloud.notifications.auth.principal.rhid.RhServiceAccountIdentity;
 import com.redhat.cloud.notifications.auth.principal.rhid.RhUserIdentity;
-import com.redhat.cloud.notifications.auth.principal.turnpike.TurnpikePrincipal;
-import com.redhat.cloud.notifications.auth.principal.turnpike.TurnpikeSamlIdentity;
 import com.redhat.cloud.notifications.config.BackendConfig;
 import io.quarkus.test.InjectMock;
 import io.quarkus.test.junit.QuarkusTest;
@@ -28,7 +26,6 @@ import org.project_kessel.api.relations.v1beta1.SubjectReference;
 import org.project_kessel.relations.client.CheckClient;
 import org.project_kessel.relations.client.LookupClient;
 
-import java.security.Principal;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
@@ -271,97 +268,6 @@ public class KesselAuthorizationTest {
 
             Assertions.assertEquals(ResourceType.INTEGRATION.getKesselName(), lookupResourcesRequest.getResourceType().getName(), String.format("unexpected resource type obtained on test case: %s", tc));
         }
-    }
-
-    /**
-     * Test that the {@link RhIdentity} is correctly extracted from a security
-     * context.
-     */
-    @Test
-    void testExtractRhIdentity() {
-        // Mock the security context.
-        final SecurityContext mockedSecurityContext = Mockito.mock(SecurityContext.class);
-
-        // Create a RhIdentity principal and assign it to the mocked security
-        // context.
-        final RhIdentity identity = Mockito.mock(RhIdentity.class);
-        Mockito.when(identity.getName()).thenReturn("Red Hat user");
-
-        final ConsolePrincipal<?> principal = new RhIdPrincipal(identity);
-        Mockito.when(mockedSecurityContext.getUserPrincipal()).thenReturn(principal);
-
-        // Call the function under test.
-        final RhIdentity result = this.kesselAuthorization.extractRhIdentity(mockedSecurityContext);
-
-        // Assert that the objects are the same. Just by checking the object's
-        // reference we can be sure that our stubbed principal above is the
-        // one that was extracted.
-        Assertions.assertEquals(
-            identity,
-            result,
-            "the extracted identity object was not the same"
-        );
-    }
-
-    /**
-     * Test that when a "non-console" principal is extracted from the security
-     * context, an exception is raised.
-     */
-    @Test
-    void testExtractRhIdentityNoConsolePrincipalThrowsException() {
-        // Mock the security context.
-        final SecurityContext mockedSecurityContext = Mockito.mock(SecurityContext.class);
-
-        // Mock a generic principal and make the context return it when asked
-        // for it.
-        final Principal mockedPrincipal = Mockito.mock(Principal.class);
-        Mockito.when(mockedSecurityContext.getUserPrincipal()).thenReturn(Mockito.mock(Principal.class));
-
-        // Call the function under test.
-        final IllegalStateException e = Assertions.assertThrows(
-            IllegalStateException.class,
-            () -> this.kesselAuthorization.extractRhIdentity(mockedSecurityContext)
-        );
-
-        // Assert that the correct exception has been thrown.
-        Assertions.assertEquals(
-            String.format("unable to extract RH Identity object from principal. Expected \"Console Principal\" object type, got \"%s\"", mockedPrincipal.getClass().getName()),
-            e.getMessage(),
-            "unexpected exception message"
-        );
-    }
-
-    /**
-     * Test that a "non-RhIdentity" identity inside a principal raises an
-     * exception.
-     */
-    @Test
-    void testExtractRhIdentityNoSupportedIdentityThrowsException() {
-        // Mock the security context.
-        final SecurityContext mockedSecurityContext = Mockito.mock(SecurityContext.class);
-
-        // Mock an unexpected identity which should trigger an exception.
-        final TurnpikeSamlIdentity turnpikeSamlIdentity = new TurnpikeSamlIdentity();
-        turnpikeSamlIdentity.associate = new TurnpikeSamlIdentity.Associate();
-        turnpikeSamlIdentity.associate.email = "example@redhat.com";
-        turnpikeSamlIdentity.type = "turnpike";
-
-        // Make the identity part of the principal.
-        final ConsolePrincipal<?> turnpikePrincipal = new TurnpikePrincipal(turnpikeSamlIdentity);
-        Mockito.when(mockedSecurityContext.getUserPrincipal()).thenReturn(turnpikePrincipal);
-
-        // Call the function under test.
-        final IllegalStateException e = Assertions.assertThrows(
-            IllegalStateException.class,
-            () -> this.kesselAuthorization.extractRhIdentity(mockedSecurityContext)
-        );
-
-        // Assert that the correct exception has been thrown.
-        Assertions.assertEquals(
-            String.format("unable to extract RH Identity object from principal. Expected \"RhIdentity\" object type, got \"%s\"", turnpikeSamlIdentity.getClass().getName()),
-            e.getMessage(),
-            "unexpected exception message"
-        );
     }
 
     /**
